# app.py
from __future__ import annotations

import os
from datetime import date, datetime
from pathlib import Path
from threading import Lock

from fastapi import FastAPI, HTTPException, Request, Response
from fastapi.exceptions import RequestValidationError
from fastapi.openapi.docs import get_swagger_ui_html
from fastapi.responses import FileResponse, HTMLResponse
from fastapi.responses import JSONResponse
from pydantic import BaseModel, Field


DATE_FORMAT = "%d.%m.%Y"


class UserPayload(BaseModel):
    name: str = Field(alias="Name")
    surname: str = Field(alias="Surname")
    date_of_birth: str = Field(alias="DateOfBirth")
    interests: list[str] = Field(alias="Interests")

    model_config = {"populate_by_name": True}


class UserResponse(BaseModel):
    user_id: int = Field(alias="Id")
    name: str = Field(alias="Name")
    surname: str = Field(alias="Surname")
    date_of_birth: str = Field(alias="DateOfBirth")
    age: int = Field(alias="Age")
    is_adult: bool = Field(alias="IsAdult")
    interests: list[str] = Field(alias="Interests")

    model_config = {"populate_by_name": True}


app = FastAPI(
    title="User management system API",
    description="API that handles user management",
    version="1.0.0",
    docs_url=None,
    redoc_url=None,
    openapi_url=None,
    swagger_ui_parameters={"defaultModelsExpandDepth": -1},
)

APP_DIR = Path(__file__).resolve().parent
SWAGGER_FILE = APP_DIR / "swagger.yaml"

_users: dict[int, dict] = {}
_next_id = 1
_lock = Lock()


def _parse_date(value: str) -> datetime:
    try:
        return datetime.strptime(value, DATE_FORMAT)
    except ValueError as exc:
        raise HTTPException(
            status_code=400,
            detail="DateOfBirth must use dd.mm.yyyy format",
        ) from exc


def _calculate_age(date_of_birth: str) -> int:
    birth_date = _parse_date(date_of_birth).date()
    today = date.today()
    return today.year - birth_date.year - (
        (today.month, today.day) < (birth_date.month, birth_date.day)
    )


def _serialize_user(user_id: int, payload: UserPayload) -> dict:
    age = _calculate_age(payload.date_of_birth)
    return {
        "Id": user_id,
        "Name": payload.name,
        "Surname": payload.surname,
        "DateOfBirth": payload.date_of_birth,
        "Age": age,
        "IsAdult": age >= 18,
        "Interests": payload.interests,
    }


def _get_user_or_404(user_id: int) -> dict:
    user = _users.get(user_id)
    if user is None:
        raise HTTPException(status_code=404, detail=f"User {user_id} not found")
    return user


@app.exception_handler(RequestValidationError)
async def request_validation_exception_handler(
    _: Request, exc: RequestValidationError
) -> JSONResponse:
    return JSONResponse(status_code=400, content={"message": exc.errors()})


@app.get("/health", include_in_schema=False)
def healthcheck() -> dict:
    return {"status": "ok"}


@app.post(
    "/api/user/create",
    summary="Create a user",
    response_model=UserResponse,
    response_model_by_alias=True,
    responses={
        200: {"description": "User created"},
        400: {"description": "Bad request"},
    },
)
async def create_user(payload: UserPayload) -> dict:
    global _next_id

    with _lock:
        user_id = _next_id
        _next_id += 1
        user = _serialize_user(user_id, payload)
        _users[user_id] = user

    return user


@app.get(
    "/api/user/id/{id}",
    summary="Get user by ID",
    response_model=UserResponse,
    response_model_by_alias=True,
    responses={
        200: {"description": "User found"},
        404: {"description": "User not found"},
    },
)
def get_user_by_id(id: int) -> dict:
    return _get_user_or_404(id)


@app.get(
    "/api/user/name/{name}",
    summary="Get users by name",
    response_model=list[UserResponse],
    response_model_by_alias=True,
    responses={200: {"description": "Success"}},
)
def get_users_by_name(name: str) -> list[dict]:
    return [
        user
        for user in _users.values()
        if user["Name"].lower() == name.lower()
    ]


@app.put(
    "/api/user/update/{id}",
    summary="Update a user by ID",
    response_model=UserResponse,
    response_model_by_alias=True,
    responses={200: {"description": "Success"}},
)
async def update_user(id: int, payload: UserPayload) -> dict:
    _get_user_or_404(id)

    with _lock:
        user = _serialize_user(id, payload)
        _users[id] = user

    return user


@app.delete(
    "/api/user/delete/{id}",
    summary="Delete a user by ID",
    status_code=200,
    responses={
        200: {"description": "Success"},
        400: {"description": "Incorrect id format"},
        404: {"description": "User not found"},
    },
)
def delete_user(id: int) -> Response:
    with _lock:
        _get_user_or_404(id)
        del _users[id]

    return Response(status_code=200)


@app.get("/swagger.yaml", include_in_schema=False)
def get_swagger_yaml() -> FileResponse:
    return FileResponse(SWAGGER_FILE, media_type="application/yaml")


@app.get("/docs", include_in_schema=False)
def custom_swagger_ui() -> HTMLResponse:
    return get_swagger_ui_html(
        openapi_url="/swagger.yaml",
        title=f"{app.title} - Swagger UI",
        swagger_ui_parameters={"defaultModelsExpandDepth": -1},
    )


if __name__ == "__main__":
    import uvicorn

    uvicorn.run("app:app", host="0.0.0.0", port=int(os.getenv("PORT", "8000")))